% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{check_onclick_}
\alias{check_onclick_}
\title{check onclick}
\usage{
check_onclick_(.menu_react, .butt_number, my_envir)
}
\value{
Change value of reactive output, without return
}
\description{
function to put a tickmark on click
}
\keyword{internal}
