% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{create_cleanfunction_}
\alias{create_cleanfunction_}
\title{create_cleanfunction_}
\usage{
create_cleanfunction_(my_envir, .which_read = c("FORWARD", "REVERSE"))
}
\value{
Function with selected cleaning operations
}
\description{
Create a function to process FASTQ files in function
of the Shiny parameters selected by the user
}
\keyword{internal}
