% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{create_uniform_width}
\alias{create_uniform_width}
\title{Create fastq/sequences/qualities with uniform width}
\usage{
create_uniform_width(input, type = c("fastq", "sequence", "quality"))
}
\arguments{
\item{input}{input to edit}

\item{type}{type of the input: 'fastq' (ShortReadQ),
'sequence' (DNAStringSet), 'quality' (BStringset)}
}
\value{
ShortReadQ object or character vector with sequences 
or qualities, with uniform widht (padded with Ns or \})
}
\description{
Create fastq/sequences/qualities with uniform width
}
\keyword{internal}
