% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_fqc.R
\name{launch_fqc}
\alias{launch_fqc}
\title{Launch FastqCleaner application}
\usage{
launch_fqc(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{Launch in browser? Default TRUE}

\item{...}{Additional parameters passed to \code{\link[shiny]{runApp}}}
}
\value{
Launch the application, without return value
}
\description{
Launch FastqCleaner application
}
\examples{
# Paste in te console to launch the application:
# launch_fqc() 

NULL
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
