% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{messageFun_}
\alias{messageFun_}
\title{messageFun_}
\usage{
messageFun_(.who, .chunck, .which_read, my_envir)
}
\value{
Changes the state of reactive vector, without return
}
\description{
messageFun_
}
\keyword{internal}
