% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjects.R
\name{myPlot}
\alias{myPlot}
\title{myplot}
\usage{
myPlot(isPaired, location, sampleSize, kmerLength, theFile, maxFreq)
}
\value{
List with Highcharts plots
}
\description{
Diagnostic plots construction
}
\keyword{internal}
