% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{outputClean_}
\alias{outputClean_}
\title{outputClean_}
\usage{
outputClean_(.myFile, .lengthWidthVec, my_envir)
}
\value{
Vector with chunks length and width  information
}
\description{
outputClean_
}
\keyword{internal}
