% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjects.R
\name{plotD}
\alias{plotD}
\title{plotD}
\usage{
plotD(x, nplots = 1, theFile = c("input", "output"), sampleSize)
}
\value{
percent of reads with quality > threshold plot
}
\description{
plotD
}
\keyword{internal}
