% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjects.R
\name{plotF}
\alias{plotF}
\title{plotF}
\usage{
plotF(x, nplots = 1, theFile = c("input", "output"), sampleSize)
}
\value{
Per cycle base proportion plot (lineplot)
}
\description{
plotF
}
\keyword{internal}
