% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjects.R
\name{plotObjects}
\alias{plotObjects}
\title{plotObjects
Create the information for the plots. This is the input of myplot}
\usage{
plotObjects(fq, klength, basename, maxFreq, sampleSize)
}
\value{
List with information to construct the diagnostic plots
}
\description{
plotObjects
Create the information for the plots. This is the input of myplot
}
\keyword{internal}
