% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{processingFunction_}
\alias{processingFunction_}
\title{processingFunction_}
\usage{
processingFunction_(my_envir)
}
\value{
Processes the input FASTQ file, without output
}
\description{
processing function for fastq files
}
\keyword{internal}
