% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{seq_names}
\alias{seq_names}
\title{Create sequences names}
\usage{
seq_names(n, base_name = "s", sep = "_")
}
\arguments{
\item{n}{Number of reads}

\item{base_name}{Base name for strings}

\item{sep}{Character separing base names and the read number. Default: '_}
}
\value{
\code{\link[Biostrings:XStringSet-class]{BStringSet}} object
}
\description{
Create \code{\link[Biostrings:XStringSet-class]{BStringSet}} 
object with names
}
\examples{
snames <- seq_names(10)
snames
snames2 <- seq_names(10, base_name = 's', sep = '.')
snames2
}
