% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjects.R
\name{myPlot}
\alias{myPlot}
\title{myPlot}
\usage{
myPlot(isPaired, location, sampleSize, kmerLength, theFile, maxFreq)
}
\value{
List with Highcharts plots
}
\description{
Construction of diagnostic plots. The function depends
of the values created by plotObject
}
\keyword{internal}
