% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjects.R
\name{plotObjects}
\alias{plotObjects}
\title{plotObjects
Create the information required to construct the plots. 
This is the input of myplot, which uses the values created for 
this function to construct the plots}
\usage{
plotObjects(fq, klength, basename, maxFreq, sampleSize)
}
\value{
List with information to construct the diagnostic plots
}
\description{
plotObjects
Create the information required to construct the plots. 
This is the input of myplot, which uses the values created for 
this function to construct the plots
}
\keyword{internal}
