% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{processingFunction_}
\alias{processingFunction_}
\title{processingFunction_}
\usage{
processingFunction_(my_envir)
}
\value{
Processes the input FASTQ file, without return
}
\description{
This function is the core of the application. It is used 
for the program to process the FASTQ file/s in the environment 
of the Shiny app. Note that this program makes a call to
create_cleanfunction
}
\keyword{internal}
